Fix
Former-commit-id: c4fe7e8a9b8c518f52d3e362d074c905d95a0eff [formerly 6e60013651d10d12b407c449006cc0ef81dff1a4]
Former-commit-id: ec362ecd130b7918c1d616f029360b94c4fee841 [formerly 2d6431259a7e60753377f52f46df38da2a1cdfb4]
Former-commit-id: cab34ab478861444909bdfccc0f174b01bf5ceca
Former-commit-id: 4b54a3946f7418015a576047f55f3d2d09256aa9

+++ Source/RawParser/Model/Parser/Decoder/Decompressor/LJpegPlain.cs
@@ -217,16 +217,16 @@ namespace RawNet.Decoder.Decompressor
             sliceWidth[slices] = sliceWidth[slices - 1];        // Extra offset to avoid branch in loop.
 
             // Predictors for components
-            int[] p = new int[4];
-            uint dest = imagePos[0];
+            long[] p = new long[4];
+            long dest = imagePos[0];
 
             // Always points to next slice
             slice = 1;
-            uint pixInSlice = sliceWidth[0];
+            long pixInSlice = sliceWidth[0];
 
             // Initialize predictors and decode one group.pitch
             uint x = 0;
-            uint predict = dest;          // Prediction pointer
+            long predict = dest;          // Prediction pointer
             for (uint i = 0; i < comps; i++)
             {
                 for (uint y2 = 0; y2 < samplesV[i]; y2++)
@@ -632,7 +632,7 @@ namespace RawNet.Decoder.Decompressor
                 UInt32 cw = frame.width - skipX;
                 for (slice = 0; slice < slices; slice++)
                 {
-                    offset[slice] = (t_x + offX+ ((offY + t_y) * raw.raw.dim.width)) | (t_s << 28);
+                    offset[slice] = (t_x + offX + ((offY + t_y) * raw.raw.dim.width)) | (t_s << 28);
                     Debug.Assert((offset[slice] & 0x0fffffff) < raw.raw.dim.width * raw.raw.dim.height);
                     t_y++;
                     if (t_y == (frame.height - skipY))
